+++ src/main/java/com/bio/dao/UserDAO.java
@@ -16,7 +16,7 @@ public interface UserDAO {
     List<User> getUsers();
 
     @Select("SELECT id,name FROM users WHERE id = #{id} and deleted_at is null")
-    User getUserbyId(@Param("id") Integer id);
+    User getUserbyId(@Param("id") Long id);
 
     @Insert("Insert into users(name) values(#{name})")
     @Options(useGeneratedKeys = true, keyColumn = "id", keyProperty = "id")
@@ -26,6 +26,6 @@ public interface UserDAO {
     void update(User user);
 
     @Delete("UPDATE users SET deleted_at = now() WHERE id = #{id}")
-    void delete(@Param("id") Integer id);
+    void delete(@Param("id") Long id);
 
 }
